+++ lib/getch/void/finalize.rb
@@ -10,10 +10,16 @@ module Getch
       protected
 
       def x
+        shell
         accounts
       end
 
       private
+
+      # Make the default shell /bin/bash instead of /bin/sh
+      def shell
+        Chroot.new('chsh -s /bin/bash')
+      end
 
       def accounts
         account = Config::Account.new